move errors into dedicated file
